backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
pydantic-settings==2.3.4
python-dotenv==1.0.1
alembic==1.13.2
google-generativeai==0.7.2
